Rename resource_spec to resource_implementation_spec and add resource_spec Refactor ResourceImplementation
